backend/requirements.txt
fastapi>=0.116,<1
uvicorn[standard]>=0.35,<1
sqlalchemy>=2.0,<3
pydantic>=2.11,<3
email-validator>=2.2,<3
python-jose[cryptography]>=3.4,<4
passlib[bcrypt]>=1.7,<2
python-multipart>=0.0.20,<1
